works

+++ IOperand.hpp
@@ -16,7 +16,7 @@ class IOperand : public Token
 {
     public:
         IOperand() {} ;
-        virtual uint64_t getVal(void) const = 0;
+        virtual double getVal(void) const = 0;
         virtual int getPrecision( void ) const = 0; // Precision of the type of the instance
         virtual eOperandType getType( void ) const = 0; // Type of the instance
         virtual IOperand const * operator+( IOperand const & rhs ) const = 0; // Sum
